Consolidate nickname into user aggregate
Nickname is an element of the user, a way (but not the only way) to
refer to that user. Consolidated into the parent aggregate.
